requirements.txt
streamlit==1.28.0
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.0.1
pandas==2.0.3